pyproject.toml
[tool.poetry]
name = "ritual-scanner"
version = "0.1.0"
description = ""
authors = ["staticf0x <staticf0x@pm.me>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "~3.12"
numpy = "^1.26.4"
opencv-python = "^4.9.0.80"
pillow = "^10.2.0"
pytesseract = "^0.3.10"
pynput = "^1.7.6"
pyperclip = "^1.8.2"
pyside6 = "6.6.2"
loguru = "^0.7.2"
rich = "^13.7.1"
scikit-image = "^0.22.0"
pydantic = "^2.6.4"


[tool.poetry.group.dev.dependencies]
devtools = {extras = ["pygments"], version = "^0.12.2"}
pytest = "^8.1.1"
pytest-cov = "^5.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.isort]
profile = "black"

[tool.ruff]
line-length = 100
target-version = "py312"

# For rule groups and specific rules, see:
# https://beta.ruff.rs/docs/rules/
select = ["ALL"]

ignore = [
    "ANN101",  # Missing type annotation for `self` in method
    "D406",  # Section name should end with a newline
    "D407",  # Missing dashed underline after section
    "D203",  # 1 blank line required before class docstring (incompatible with D211)
    "D213",  # Multi-line docstring summary should start at the second line (incompatible with D212)
    "D105",  # Missing docstring in magic method
    "D107",  # Missing docstring in `__init__`
    "TD",
]

[tool.ruff.mccabe]
max-complexity = 10

[tool.ruff.per-file-ignores]
"tests/*" = [
    "ANN",  # All type annotations
    "D100",  # Missing docstring in public module
    "D101",  # Missing docstring in public class
    "D102",  # Missing docstring in public method
    "D103",  # Missing docstring in public function
    "S101",  # Use of `assert` detected
    "PLR2004",  # Magic value used in comparison
    "SLF001", # Private member access
]
"scanner/constants.py" = [
    "ERA001",
]
"items.py" = ["T201", "SIM115"]
"main.py" = ["S701"]
"benchmark.py" = ["T201"]

[[tool.mypy.overrides]]
module = "pytesseract"
ignore_missing_imports = true
